+++ packages/grid/src/features/table-ui/components/table/TableRow.tsx
@@ -1,16 +1,22 @@
 import { BoxProps, Row } from "@stenajs-webui/core";
 import * as React from "react";
-import { tableBorder, tableRowHeight } from "../../../../config/TableConfig";
+import {
+  defaultTableHeadRowHeight,
+  tableBorder,
+} from "../../../../config/TableConfig";
 
 interface Props extends BoxProps {}
 
-export const TableRow: React.FC<Props> = (props) => {
+export const TableRow: React.FC<Props> = ({
+  height = defaultTableHeadRowHeight,
+  ...boxProps
+}) => {
   return (
     <Row
-      height={tableRowHeight}
       borderTop={tableBorder}
       hoverBackground={"var(--lhds-color-ui-100)"}
-      {...props}
+      height={height}
+      {...boxProps}
     />
   );
 };